ci: shippable: add secondary cache server
Shippable has a maximum build time of 1 hour for Open Source projects.
It is usually sufficient since typical build time is currently around
35 minutes, thanks to caching (ccache). But it can happen that the code
is changed too much by a pull request in which case the cache is
ineffective and build time would exceed 1 hour. When a build times out,
the cache is NOT updated and there is no other choice than to manually
feed the cache with a partial build (by removing some steps in
.shippable.yml) so that the next build can be quicker, wait for the
build to succeed, then restore all the build steps. This is quite
impractical.

This commit introduces a secondary cache on a private server with
automatic upload after 50 minutes of build time (or at the end of the
build). Then if the build times out we still have cached a significant
amount of data so restarting the build should be enough to deal with
the problem.

If the server is unavailable for any reason, errors are ignored and the
build runs as before using the cache stored in the Shippable
infrastructure.

Cache data are transferred using SSH and the cache server is configured
with the public key of the official OP-TEE OS Shippable project (the
one that is used to build pull requests). Each Shippable project has
its own SSH keypair, so forks of the main project can't use the server
by default.

Signed-off-by: Jerome Forissier <jerome@forissier.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>