core, ldelf: add gdb helper functions
<fixup commit, intention is to use this update commit message when squashing>

When running trying to debug Trusted Applications with GDB one must
figure out the load adress to be able to load the symbols for the TA at
the correct offset. The simple way has been to boot up the and run the
TA once and then look for the load address in UART log. Even though that
works, it's error prone and cumbersome. It's error prone because it's no
guarantee that the load address will stay the same and if you run a use
case that involves more than a single TA, then it's hard to know the
(load order) of load addresses. If you run with ASLR enabled for TA's,
then the load address will definitely change, so there it's impossible
to use the UART output.

From GDB it's possible to automate finding of TA load addresses by
setting strategic breakpoints and parse the stack frames and parameters.
However, that involve quite a bit of work and it is also a bit error
prone.

By adding helper functions like in this patch one can from GDB in a
pretty easy manner grab the correct load address, by using the
parameters in the helper functions. This is something that also works
when compiling TA's with ASLR enabled.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>