libtomcrypt: se050: pick and chose operations
For performance reasons it is convenient to execute certain crypto
operations in the processor rather than routing it to a secure element
over i2c.

Other operations, like those related to SCP03 must be executed on the
processor; these would be operations to encrypt/decrypt/validate the
data in the i2c buffers.

Signed-off-by: Jorge Ramirez-Ortiz <jorge@foundries.io>