scripts: add remote processor firmware signature tool
Add a script that signs one or several remote processor ELF firmware that
will be authenticated by the remoteproc TA.

This tool adds a binary header, a signature and a TLV list.

The header contains a magic number, a version number and
the size of the different blobs (signature, images, TLV list blobs).

The signature contains a signature authenticating the
header blob hash and the TLV blob hash.

The TLV blob contains a list of data formatted as Type/Length/Value
fields. It contains information for the remoteproc TA and
the remoteproc platform specific PTA.

The TLV types from 0 to 0x00010000 are predefined information used by
the remoteproc TA:
- algorithm used for signature
- algorithm used for computing segment's hash
- number of images to load
- types of the images to load
- sizes of the images to load
- a copy of the elf segment tables with associated hash

the TLV types from 0x00010000 to 0x00020000 contains information
transferred to the remoteproc platform PTA.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>
Reviewed-by: Etienne Carriere <etienne.carriere@foss.st.com>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>