libgcov: Add library adding the gcov support
The code coverage is a metric measuring the lines of code
executed in a code base during execution.
If the code coverage is generated after a test suite is
ran, it will give the coverage of code tested by the
test suite. It gives the possibility to identify the code
which is not tested.

The gcov code coverage is supported by different compiler:
 - GCC (gcc.gnu.org/onlinedocs/gcc/Gcov.html)
 - clang compiler (clang.llvm.org/docs/SourceBasedCodeCoverage.html)
It allows to analyze which part of the code is executed.
The interface with a specific implementation of gcov is defined
usign the structure gcov_impl_t.

The library implements the symbol required for gcov:
 - __gcov_init: called by each coverage unit
 - __gcov_merge_add
 - __gcov_exit
The calls to these symbols are redirected to the implementation
for the compiler used.

The library role is to transform the coverage data in program
memory to a valid gcov data file. For this it interacts with:
 - user/TA: Control (Reset or Dump) the coverage data
 - Function to store data: Send the data to it for storage in REE FS

In order to retrieve the coverage (*.gcda) for a specific UC,
it is recommended to do the following actions:
 - clear the coverage
 - perform the UC
 - save the coverage

Note: The core and the TAs are linked against different instances
of the libgcov. The libgcov of the core can be controlled by the
gcov PTA. The TA instances are controlled calling the libgcov API.

Signed-off-by: Franck LENORMAND <franck.lenormand@nxp.com>