coverage.py: Add script to process code coverage
This script ease the generation of code coverage report. A report
is based on a codebase, note files (generated at compilation) and
data files (generated at runtime).
The report is generated using lcov and present the code coverage as
a percentage of lines executed compared to the total number of lines
in the codebase.

The script will copy all the note and data files in a temporary
directory allowing to reuse the generated artifact and checks that
the note and data file are correctly organised (the files must be
in the same directory).

The script also allow to generate an html output using genhtml and
to view it.

Signed-off-by: Franck LENORMAND <franck.lenormand@nxp.com>