core: mm: move boot_mmu_config from .text to .bss
The boot_mmu_config struct is currently in the .text section and its
content is written to during boot in core_init_mmu_map(). As a result,
the .text section at run time is different from what it is in tee.elf.
This is not a good thing and it makes matters more complicated than
necessary for instance when implementing code integrity checks.

Move the struct definition to C code (and remove the wrong 'const'
qualifier in the declaration) so that the symbol is allocated in the
.bss section. Access from a32 assembly needs to be changed in a couple
of places to avoid assembler errors ('adr/ldr Rd, boot_mmu_config').
The same technique as for stack_tmp_export and stack_tmp_stride is used.

Signed-off-by: Jerome Forissier <jerome@forissier.org>