drivers: imx_i2c: add support for imx6ull
Support native I2C access on imx6ull (this SoC has an additional
register - compared to the imx8mm - to configure the daisy chain in
the iomuxc).

The driver allows the user to specify the clock rate even though it
should have been configured to use the 24MHz oscillator as recommended
by the I2C errata [1]

A patch has been sent to U-boot to address their current release as of
Oct 23, 2020 where the peripheral clock is still set to 66MHz.

[1] imx6ull errata, Rev 2, 10/2019, ERR007805
Tested on imx6ull-evk 14x14 with the bus at 100Kbps.

Signed-off-by: Jorge Ramirez-Ortiz <jorge@foundries.io>