core: Add support for code coverage compilation
The code coverage is performed at runtime by keeping in memory
structures holding the releveant data:

In order to generate the code coverage, two steps are required:
 - Instruct the compiler to generate coverage information structure
   and functions to initialize them: flag --coverage.
 - Call the constructors added to initialize the coverage information
   and register them to the libgcov (modification of boot.c)

When an object file (.o) is created with code coverage support, a note
(.gcno) file is created with it. The note file contains informations
used when creating the code coverage report.

The support of code coverage in compiled code increases its size (~15%)
and the compilation time (~25%).

Signed-off-by: Franck LENORMAND <franck.lenormand@nxp.com>